+++ eslint.config.js
@@ -1,14 +1,21 @@
 import js from '@eslint/js';
 import reactPlugin from 'eslint-plugin-react';
+import tseslint from '@typescript-eslint/eslint-plugin';
+import tsParser from '@typescript-eslint/parser';
 
 export default [
   js.configs.recommended,
   {
     files: ['**/*.{ts,tsx}'],
-    plugins: { react: reactPlugin },
+    plugins: {
+      react: reactPlugin,
+      '@typescript-eslint': tseslint,
+    },
     languageOptions: {
+      parser: tsParser,
       parserOptions: {
         ecmaFeatures: { jsx: true },
+        sourceType: 'module',
       },
     },
     settings: {
@@ -16,6 +23,9 @@ export default [
     },
     rules: {
       'react/react-in-jsx-scope': 'off',
+      'no-unused-vars': 'off',
+      '@typescript-eslint/no-unused-vars': ['warn', { argsIgnorePattern: '^_' }],
+      'no-undef': 'off', // TypeScript handles this
     },
   },
   {